p3-worker/requirements.txt
boto3==1.40.25
python-dotenv==1.1.1